feat: add g3-5 content for Skyshield
